Allow IPT scan to use API key for authentication.
API Key is supplied via a cookie to ensure key is not part of the URL and since the scan is a plain get without a payload. See IptController for documentation.

API Key requires the following configuration settings in ala-collectory-config.properties which are not part of the standard defaults:

security.apikey.checkEnabled=true
security.apikey.serviceUrl=http://auth.ala.org.au/apikey/ws/check?apikey=